途中経過

+++ Make_Draw_LineContainer.h
@@ -32,14 +32,14 @@ namespace Make {
 
 			const std::uint16_t barID; //何小節目に属しているか
 			const std::uint8_t beatID;//その小節の何番目の線か
-			const float time;//曲の開始から何秒か
+			const double time;//曲の開始から何秒か
 			const std::uint8_t laneAmount; //レーンの数
-			float yMax;//座標の最大値
-			float yMin;//座標の最小値
+			double yMax;//座標の最大値
+			double yMin;//座標の最小値
 			std::uint32_t color;//拍線の色
 			std::uint8_t lineThickness; //拍線の太さ
-			float y; //拍線の座標
-			std::vector<float> laneX;
+			double y; //拍線の座標
+			std::vector<double> laneX;
 			std::uint8_t barIDThickness; //小節番号の線の太さ
 			std::uint32_t barIDColor;//小節番号の色
 			std::uint8_t barIDStrWidth;//小節番号の文字の大きさ
@@ -51,20 +51,20 @@ namespace Make {
 			void drawLine() ;
 			void drawBarID() ;
 		public:
-			Make_Draw_LineContainer(const std::uint16_t barID,const float time,const std::uint8_t beatID,const float y,const float yMax,const std::shared_ptr<Note::Make_Note_NoteManager>& p_noteManager);
+			Make_Draw_LineContainer(const std::uint16_t barID,const double time,const std::uint8_t beatID,const double y,const double yMax,const std::shared_ptr<Note::Make_Note_NoteManager>& p_noteManager);
 			static void setNoteType(const std::uint8_t type) ;
 			static std::uint16_t getbarIDForChangeQuontize();
 
-			void setYMin(const float y);
-			void updateYMax(const float y);
-			void updateByInitOneBar(const float yWidth);
-			void updateY(const float y);
+			void setYMin(const double y);
+			void updateYMax(const double y);
+			void updateByInitOneBar(const double yWidth);
+			void updateY(const double y);
 			void draw();
 
-			float getTime();
-			float getY();
-			float getYMin();
-			float getYMax();
+			double getTime();
+			double getY();
+			double getYMin();
+			double getYMax();
 		};
 	}
 }
