+++ Make_Note_NormalNoteContainer.cpp
@@ -1,27 +1,28 @@
 #include "Make_Note_NormalNoteContainer.h"
 
 
-float Make::Note::Make_Note_NormalNoteContainer::noteWidth = 15.0;
+double Make::Note::Make_Note_NormalNoteContainer::noteWidth = 15.0;
 
-Make::Note::Make_Note_NormalNoteContainer::Make_Note_NormalNoteContainer(const std::uint16_t barID,const std::uint8_t beatID,const float& y,const std::uint8_t laneAmount,const float time) :
+Make::Note::Make_Note_NormalNoteContainer::Make_Note_NormalNoteContainer(const std::uint16_t barID,const std::uint8_t beatID,const double& y,const std::uint8_t laneAmount,const double time) :
 	barID(barID),beatID(beatID),laneAmount(laneAmount),time(time),r_y(y) {
 	noteX.resize(laneAmount);
 	noteFlag.resize(laneAmount);
-	float laneWidth = (Global::DRAW_X_MAX - Global::DRAW_X_MIN) / laneAmount;
+	double laneWidth = (Global::DRAW_X_MAX - Global::DRAW_X_MIN) / laneAmount;
 	for (int i = 0; i < laneAmount; ++i) {
-		noteX.at(i) = laneWidth * i + Global::DRAW_X_MIN + laneWidth * 0.5f;
+		noteX.at(i) = laneWidth * i + Global::DRAW_X_MIN + laneWidth * 0.5;
 		noteFlag.at(i) = false;
 	}
 	color = GetColor(255, 255, 255);
-	notePointX = noteWidth * 1.2f;
-	notePointY = noteWidth * 0.3f;
+	notePointX = noteWidth * 1.2;
+	notePointY = noteWidth * 0.3;
 }
 
 void Make::Note::Make_Note_NormalNoteContainer::drawNote() {
 	if (r_y < Make::Global::WINDOW_HEIGHT && r_y>0) {
 		for (int i = 0, iSize = static_cast<int>(noteFlag.size()); i < iSize; ++i) {
 			if (noteFlag.at(i)) {
-				DrawBoxAA(noteX.at(i) - notePointX, r_y - notePointY, noteX.at(i) + notePointX, r_y + notePointY, color, true);
+				DrawBoxAA(static_cast<float>(noteX.at(i) - notePointX), static_cast<float>(r_y - notePointY),
+					static_cast<float>(noteX.at(i) + notePointX), static_cast<float>(r_y + notePointY), color, true);
 			}
 		}
 	}
@@ -40,7 +41,7 @@ const bool Make::Note::Make_Note_NormalNoteContainer::getNormalNoteFlag(const st
 	return noteFlag.at(laneID);
 }
 
-const float& Make::Note::Make_Note_NormalNoteContainer::getTime() {
+const double& Make::Note::Make_Note_NormalNoteContainer::getTime() {
 	return time;
 }
